Implementation of locally nameless de-Bruijn vars.
Var and var types now refer to names only in the
initial parse.  `numberAst` replaces variable
names with de-Bruijn indices.  The stack form now
looks up those numbers -or- uses pointers directly.
The second form (pointers) occurs when get_ast
is called on a sub-tree.  Variables inside the sub-tree
are numbers, and variables outside the sub-tree are
pointers.

This fixes the by-need evaluation in principle.
Testing is required.