Fixed numberAst, added groups.
numberAst replaced variable numbers in-place, which
caused conflicts when Ast-s shared var-nodes.
This version creates fresh Ast-trees as it traverses.
Group/group is implemented in the ast, but not the
stack form.
The pretty-printer has also been improved.
Basic eval by-need is working.
Note that evaluating applications under lambda leads to
trouble with the if-statement when no boolean
is provided (since both true and false
branches are evaluated).